[FIX] double para big decimal

+++ backend/src/main/java/br/com/financialsys/backend/dto/BankDTO.java
@@ -1,5 +1,7 @@
 package br.com.financialsys.backend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,5 +14,5 @@ import lombok.Setter;
 public class BankDTO {
     private Long idBank;
     private String name;
-    private double initialBalance;
+    private BigDecimal initialBalance;
 }
